* 3.0-dev: (75 commits)
  创建评论处理  emoji 等
  站点信息接口修改
  打印京东抓取数据
  管理端数据面板   网站配置参数添加
  添加账号密码注册、手机号注册两种模式的注册原因
  话题列表控制threads返回
  修改背景图，上传头像权限优化
  草稿帖子查看权限调整
  删除调试代码
  手机号注册绑定微信时使用微信昵称更新用户昵称
  在浏览器头部伪造user-agent
  修改红包贴无法领取的问题
  修复评论内容太长导致截断乱码，不能展示评论列表的问题
  生成access_token时密码需要为空
  屏蔽附件单独付费代码
  修复集赞领红包，达到集赞要求数没有获取到红包
  完善 empty 判断
  商品帖爬取数据问题,当前用户编辑帖子返回帖子用户信息问题
  PC绑定微信的轮询接口增加返还登录态信息
  完善一下
  ...

# Conflicts:
#	app/Api/Controller/UsersV3/UploadAvatarController.php
#	app/Models/ThreadVideo.php

+++ app/Commands/Trade/PayOrder.php
@@ -172,7 +172,7 @@ class PayOrder
         }
 
         // 正确后清除错误记录
-        if ($failCount > 0) {
+        if ($this->data->get('payment_type') == Order::PAYMENT_TYPE_WALLET && $failCount > 0) {
             UserWalletFailLogs::deleteAll($this->actor->id);
         }
